+++ lib/visitors.js
@@ -1,6 +1,4 @@
-const babylon = require('babylon')
-const codeWalk = require('babel-traverse').default
-const pushUniqueVariable = require('./pushUniqueVariable')
+const parseAndFindVarsInProgram = require('./parseAndFindVarsInProgram')
 
 function isComponent(string) {
   return /^[A-Z].*$/.test(string)
@@ -8,31 +6,6 @@ function isComponent(string) {
 
 function isSpreadOperator(string) {
   return /^\.{3}[A-z]+$/.test(string)
-}
-
-function parseAndFindVarsInProgram(program) {
-  let usedVariables = []
-  let definedVariables = []
-  const code = babylon.parse(program, {
-    plugins: [
-      'objectRestSpread',
-    ],
-  })
-
-  codeWalk(code, {
-    Program(path) {
-      const varsDefined = Object.keys(path.scope.bindings)
-      const varsInGlobal = Object.keys(path.scope.globals)
-
-      usedVariables = pushUniqueVariable(usedVariables, ...varsInGlobal)
-      definedVariables = pushUniqueVariable(definedVariables, ...varsDefined)
-    },
-  })
-
-  return {
-    used: usedVariables,
-    defined: definedVariables,
-  }
 }
 
 module.exports = function visitors(state, token) {
